:star2: feat/impr(commit-tool): improve body editor
When editing using a terminal editor, it could sometimes cause weird
behavior. That's why, when available, you can now edit the commit
message using vscode.
