+++ packages/onezero-themes/package.json
@@ -24,6 +24,9 @@
     "email": "code-support@onezero.company",
     "url": "https://onezero.company"
   },
+  "scripts": {
+    "build": "echo 'This is a theme package and does not need to be built.'"
+  },
   "contributes": {
     "themes": [
       {
